+++ example/category_keyword/category_keyword_category_test.dart
@@ -55,10 +55,10 @@ void testCategoryKeywordCategory(
       var added = categories.add(category); 
       expect(added, isFalse); 
       expect(categories.length, equals(categoryCount)); 
-      expect(categories.errors.length, greaterThan(0)); 
-      expect(categories.errors.toList()[0].category, equals("required")); 
+      expect(categories.exceptions.length, greaterThan(0)); 
+      expect(categories.exceptions.toList()[0].category, equals("required")); 
  
-      categories.errors.display(title: "Add category required error"); 
+      categories.exceptions.display(title: "Add category required error"); 
     }); 
  
     test("Add category unique error", () { 
@@ -70,9 +70,9 @@ void testCategoryKeywordCategory(
       var added = categories.add(category); 
       expect(added, isFalse); 
       expect(categories.length, equals(categoryCount)); 
-      expect(categories.errors.length, greaterThan(0)); 
+      expect(categories.exceptions.length, greaterThan(0)); 
  
-      categories.errors.display(title: "Add category unique error"); 
+      categories.exceptions.display(title: "Add category unique error"); 
     }); 
  
     test("Find category by name", () {