+++ example/category_links/category_links_category_test.dart
@@ -23,7 +23,7 @@ void testCategoryLinksCategories(
       linksModel.clear(); 
       expect(linksModel.isEmpty, isTrue); 
       expect(categories.isEmpty, isTrue); 
-      expect(categories.errors.isEmpty, isTrue); 
+      expect(categories.exceptions.isEmpty, isTrue); 
     }); 
  
     test("From model to JSON", () { 
@@ -72,10 +72,10 @@ void testCategoryLinksCategories(
       var added = categories.add(category); 
       expect(added, isFalse); 
       expect(categories.length, equals(categoryCount)); 
-      expect(categories.errors.length, greaterThan(0)); 
-      expect(categories.errors.toList()[0].category, equals("required")); 
+      expect(categories.exceptions.length, greaterThan(0)); 
+      expect(categories.exceptions.toList()[0].category, equals("required")); 
  
-      categories.errors.display(title: "Add category required error"); 
+      categories.exceptions.display(title: "Add category required error"); 
     }); 
  
     test("Add category unique error", () { 
@@ -87,9 +87,9 @@ void testCategoryLinksCategories(
       var added = categories.add(category); 
       expect(added, isFalse); 
       expect(categories.length, equals(categoryCount)); 
-      expect(categories.errors.length, greaterThan(0)); 
+      expect(categories.exceptions.length, greaterThan(0)); 
  
-      categories.errors.display(title: "Add category unique error"); 
+      categories.exceptions.display(title: "Add category unique error"); 
     }); 
  
     test("Not found category by new oid", () {