+++ example/category_links/category_links_comment0_test.dart
@@ -55,10 +55,10 @@ void testCategoryLinksComment(
       var added = comments.add(comment); 
       expect(added, isFalse); 
       expect(comments.length, equals(commentCount)); 
-      expect(comments.errors.length, greaterThan(0)); 
-      expect(comments.errors.toList()[0].category, equals("required")); 
+      expect(comments.exceptions.length, greaterThan(0)); 
+      expect(comments.exceptions.toList()[0].category, equals("required")); 
  
-      comments.errors.display(title: "Add comment required error"); 
+      comments.exceptions.display(title: "Add comment required error"); 
     }); 
  
     test("Add comment unique error", () {