+++ example/category_links/category_links_member0_test.dart
@@ -55,10 +55,10 @@ void testCategoryLinksMember(
       var added = members.add(member); 
       expect(added, isFalse); 
       expect(members.length, equals(memberCount)); 
-      expect(members.errors.length, greaterThan(0)); 
-      expect(members.errors.toList()[0].category, equals("required")); 
+      expect(members.exceptions.length, greaterThan(0)); 
+      expect(members.exceptions.toList()[0].category, equals("required")); 
  
-      members.errors.display(title: "Add member required error"); 
+      members.exceptions.display(title: "Add member required error"); 
     }); 
  
     test("Add member unique error", () { 
@@ -70,9 +70,9 @@ void testCategoryLinksMember(
       var added = members.add(member); 
       expect(added, isFalse); 
       expect(members.length, equals(memberCount)); 
-      expect(members.errors.length, greaterThan(0)); 
+      expect(members.exceptions.length, greaterThan(0)); 
  
-      members.errors.display(title: "Add member unique error"); 
+      members.exceptions.display(title: "Add member unique error"); 
     }); 
  
     test("Find member by firstName", () {