+++ example/category_links/category_links_member_test.dart
@@ -23,7 +23,7 @@ void testCategoryLinksMembers(
       linksModel.clear(); 
       expect(linksModel.isEmpty, isTrue); 
       expect(members.isEmpty, isTrue); 
-      expect(members.errors.isEmpty, isTrue); 
+      expect(members.exceptions.isEmpty, isTrue); 
     }); 
  
     test("From model to JSON", () { 
@@ -72,10 +72,10 @@ void testCategoryLinksMembers(
       var added = members.add(member); 
       expect(added, isFalse); 
       expect(members.length, equals(memberCount)); 
-      expect(members.errors.length, greaterThan(0)); 
-      expect(members.errors.toList()[0].category, equals("required")); 
+      expect(members.exceptions.length, greaterThan(0)); 
+      expect(members.exceptions.toList()[0].category, equals("required")); 
  
-      members.errors.display(title: "Add member required error"); 
+      members.exceptions.display(title: "Add member required error"); 
     }); 
  
     test("Add member unique error", () { 
@@ -87,9 +87,9 @@ void testCategoryLinksMembers(
       var added = members.add(member); 
       expect(added, isFalse); 
       expect(members.length, equals(memberCount)); 
-      expect(members.errors.length, greaterThan(0)); 
+      expect(members.exceptions.length, greaterThan(0)); 
  
-      members.errors.display(title: "Add member unique error"); 
+      members.exceptions.display(title: "Add member unique error"); 
     }); 
  
     test("Not found member by new oid", () {