+++ example/category_links/category_links_question0_test.dart
@@ -55,10 +55,10 @@ void testCategoryLinksQuestion(
       var added = questions.add(question); 
       expect(added, isFalse); 
       expect(questions.length, equals(questionCount)); 
-      expect(questions.errors.length, greaterThan(0)); 
-      expect(questions.errors.toList()[0].category, equals("required")); 
+      expect(questions.exceptions.length, greaterThan(0)); 
+      expect(questions.exceptions.toList()[0].category, equals("required")); 
  
-      questions.errors.display(title: "Add question required error"); 
+      questions.exceptions.display(title: "Add question required error"); 
     }); 
  
     test("Add question unique error", () {