+++ example/dartling_todos/todo/mvc/tasks.dart
@@ -29,7 +29,7 @@ class Tasks extends TasksGen {
   Tasks(Concept concept) : super(concept);
 
   // begin: added by hand
-  Tasks get completed => selectWhere((task) => task.isCompleted);
+  Tasks get completed => selectWhere((task) => task.completed);
   Tasks get left => selectWhere((task) => task.left);
 
   bool preAdd(Task task) {
@@ -37,11 +37,11 @@ class Tasks extends TasksGen {
     if (validation) {
       validation = task.title.length <= 64;
       if (!validation) {
-        var error = new ValidationError('pre');
+        var error = new ValidationException('pre');
         error.message =
             '${concept.codePlural}.preAdd rejects the "${task.title}" title, '
             'because it is longer than 64.';
-        errors.add(error);
+        exceptions.add(error);
       }
     }
     return validation;